+++ src/org/jpc/emulator/execution/opcodes/vm/xor_Ed_Gd_mem.java
@@ -50,8 +50,8 @@ public class xor_Ed_Gd_mem extends Executable
     {
         Reg op2 = cpu.regs[op2Index];
         cpu.of = cpu.af = cpu.cf = false;
-        cpu.flagResult = (short)(op1.get32(cpu) ^ op2.get32());
-        op1.set32(cpu, (short)cpu.flagResult);
+        cpu.flagResult = (op1.get32(cpu) ^ op2.get32());
+        op1.set32(cpu, cpu.flagResult);
         cpu.flagStatus = SZP;
         return Branch.None;
     }
